refactor: re-design simple mode toggle.
